krishi-sakhi/backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
pydantic==2.6.1
SQLAlchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1
python-dotenv==1.0.1
requests==2.31.0
openai==1.14.2
redis==5.0.1
